Fixed stage resize policy for osx backend
Added initialization of minimum window size property on Cocoa
side. This property works when user change window size by mouse
dragging. But when size is changed by clutter_actor_set_size this
property will not help and was added another check in clutter_stage_osx_resize.
Also osx_get_geometry was refactoried because it returns incorrect values in many cases but now
size is saved in [Window reshape] in requisition_width/height and this
value will be returned in any case to frontend.